Restructure core code into library "Core"
