+++ lib/Core/Auth.php
