+++ lib/Core/Email.php
